refactor(frontend): split api/index.js into transport + session + domain modules
Closes Wave 5 scope doc row D1. Splits the 1009-LOC monolith at
frontend/src/api/index.js into:

  - transport.js  — axios instance + refresh-queue state (co-located
                    per scope doc §5 R5 mitigation 1)
  - session.js    — localStorage token storage (no axios dependency)
  - domain/*.js   — 11 files, one per section marker in the original:
                    phenopackets, aggregations, admin, publications,
                    auth, hpo, clinical, variants, reference,
                    variant_annotation, search
  - index.js      — rewritten as ≤100 LOC re-export aggregator preserving
                    both named exports AND the legacy default export shape

Pure refactor: no function renamed, no signature changed, no body
refactored. Every existing import site continues to work unchanged.

New transport.spec.js covers the refresh-queue thunder-herd guard.
New session.spec.js covers token storage contract.
