+++ frontend/src/api/index.js
@@ -1,1009 +1,46 @@
-// src/api/index.js - Complete rewrite for v2 API (GA4GH Phenopackets)
-import axios from 'axios';
+// src/api/index.js — Re-export aggregator
+// Preserves both named exports AND the legacy `export default` shape.
+
+export { apiClient } from './transport';
+export { getAccessToken, getRefreshToken, persistTokens, clearTokens } from './session';
+
+export * from './domain/phenopackets';
+export * from './domain/aggregations';
+export * from './domain/admin';
+export * from './domain/publications';
+export * from './domain/auth';
+export * from './domain/hpo';
+export * from './domain/clinical';
+export * from './domain/variants';
+export * from './domain/reference';
+export * from './domain/variant_annotation';
+export * from './domain/search';
+
+// Legacy default export — used by `import API from '@/api'`
+import { apiClient } from './transport';
+import * as phenopackets from './domain/phenopackets';
+import * as aggregations from './domain/aggregations';
+import * as admin from './domain/admin';
+import * as publications from './domain/publications';
+import * as auth from './domain/auth';
+import * as hpo from './domain/hpo';
+import * as clinical from './domain/clinical';
+import * as variants from './domain/variants';
+import * as reference from './domain/reference';
+import * as variantAnnotation from './domain/variant_annotation';
+import * as search from './domain/search';
 
-export const apiClient = axios.create({
-  // Use Vite proxy in development (avoids CORS), direct URL in production
-  baseURL: import.meta.env.VITE_API_URL || '/api/v2',
-  timeout: 10000,
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
-
-// Flag to prevent infinite refresh loops
-let isRefreshing = false;
-// Queue for failed requests during token refresh
-let failedRequestsQueue = [];
-
-/**
- * Process queued requests after token refresh.
- * @param {Error|null} error - Error if refresh failed
- * @param {string|null} token - New access token if refresh succeeded
- */
-function processQueue(error, token = null) {
-  failedRequestsQueue.forEach((promise) => {
-    if (error) {
-      promise.reject(error);
-    } else {
-      promise.resolve(token);
-    }
-  });
-  failedRequestsQueue = [];
-}
-
-// Request interceptor: Add JWT token from localStorage
-// Note: We use localStorage directly here for performance (synchronous)
-// The auth store manages the same tokens and keeps them in sync
-apiClient.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('access_token');
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => Promise.reject(error)
-);
-
-// Response interceptor: Handle token refresh on 401 errors
-apiClient.interceptors.response.use(
-  (response) => response,
-  async (error) => {
-    const originalRequest = error.config;
-
-    // Normalize standardized backend error shape (Wave 2 T11):
-    //   { detail, error_code, request_id }
-    // Backwards-compatible fallback for legacy FastAPI errors which only
-    // had { detail } (string or array). Attaches `error.normalized` so
-    // downstream callers can read a uniform shape.
-    const responseData = error.response?.data;
-    let normalizedDetail = error.message;
-    if (responseData) {
-      if (typeof responseData.detail === 'string') {
-        normalizedDetail = responseData.detail;
-      } else if (Array.isArray(responseData.detail)) {
-        // Legacy FastAPI validation error shape (list of {loc, msg, type})
-        normalizedDetail = responseData.detail
-          .map((item) =>
-            item && typeof item === 'object' && 'msg' in item ? item.msg : String(item)
-          )
-          .join('; ');
-      } else if (responseData.detail && typeof responseData.detail === 'object') {
-        // Structured dict details preserved by the backend handler (e.g.,
-        // {error: "conflict", current_revision: 5} from the update-conflict
-        // endpoint). Prefer a human-readable message field if present; fall
-        // back to JSON serialization so diagnostics aren't lost as
-        // "[object Object]".
-        const d = responseData.detail;
-        if (typeof d.message === 'string') {
-          normalizedDetail = d.message;
-        } else if (typeof d.error === 'string') {
-          normalizedDetail = d.error;
-        } else {
-          try {
-            normalizedDetail = JSON.stringify(d);
-          } catch {
-            normalizedDetail = String(d);
-          }
-        }
-      } else if (responseData.detail != null) {
-        normalizedDetail = String(responseData.detail);
-      }
-    }
-    error.normalized = {
-      detail: normalizedDetail,
-      errorCode: responseData?.error_code ?? null,
-      requestId: responseData?.request_id ?? null,
-    };
-
-    // Log normalized error (logService redacts sensitive fields automatically)
-    if (window.logService && error.response) {
-      window.logService.error('API request failed', {
-        status: error.response.status,
-        url: originalRequest?.url,
-        detail: error.normalized.detail,
-        errorCode: error.normalized.errorCode,
-        requestId: error.normalized.requestId,
-      });
-    }
-
-    // If 401 and not already retrying, attempt token refresh
-    if (error.response?.status === 401 && !originalRequest._retry) {
-      // Skip refresh for auth endpoints to prevent infinite loops
-      if (
-        originalRequest.url?.includes('/auth/login') ||
-        originalRequest.url?.includes('/auth/refresh')
-      ) {
-        return Promise.reject(error);
-      }
-
-      if (isRefreshing) {
-        // Another request is already refreshing, queue this one
-        return new Promise((resolve, reject) => {
-          failedRequestsQueue.push({ resolve, reject });
-        })
-          .then((token) => {
-            originalRequest.headers.Authorization = `Bearer ${token}`;
-            return apiClient(originalRequest);
-          })
-          .catch((err) => Promise.reject(err));
-      }
-
-      originalRequest._retry = true;
-      isRefreshing = true;
-
-      try {
-        // Import auth store dynamically to avoid circular dependency
-        const { useAuthStore } = await import('@/stores/authStore');
-        const authStore = useAuthStore();
-
-        // Attempt to refresh access token
-        const newAccessToken = await authStore.refreshAccessToken();
-
-        // Success! Process queued requests with new token
-        processQueue(null, newAccessToken);
-        isRefreshing = false;
-
-        // Retry original request with new token
-        originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
-        return apiClient(originalRequest);
-      } catch (refreshError) {
-        // Refresh failed - clear queue and redirect to login
-        processQueue(refreshError, null);
-        isRefreshing = false;
-
-        // Clear auth state and redirect
-        window.logService.warn('Token refresh failed, redirecting to login');
-        localStorage.removeItem('access_token');
-        localStorage.removeItem('refresh_token');
-
-        // Only redirect if not already on login page
-        if (window.location.pathname !== '/login') {
-          window.location.href = '/login';
-        }
-
-        return Promise.reject(refreshError);
-      }
-    }
-
-    // For other errors or if retry failed, just reject
-    return Promise.reject(error);
-  }
-);
-
-/**
- * DEPRECATED: Use cursor pagination instead.
- * This function is kept for backwards compatibility during migration.
- * @deprecated Use buildCursorParams from @/utils/pagination instead
- */
-export function pageToSkipLimit(page, pageSize) {
-  return {
-    skip: (page - 1) * pageSize,
-    limit: pageSize,
-  };
-}
-
-/* ==================== PHENOPACKETS ENDPOINTS ==================== */
-
-/**
- * Get a list of phenopackets with JSON:API pagination and filtering.
- * Supports offset pagination, cursor pagination, and legacy skip/limit for backwards compatibility.
- *
- * **JSON:API Offset Pagination (Simple):**
- * @param {Object} params - Query parameters
- *   - page[number]: Page number (1-indexed, default: 1)
- *   - page[size]: Items per page (default: 100, max: 500)
- *   - filter[sex]: Filter by sex (MALE, FEMALE, OTHER_SEX, UNKNOWN_SEX)
- *   - filter[has_variants]: Filter by variant presence (boolean)
- *   - sort: Comma-separated fields to sort by (prefix with '-' for descending)
- *
- * **JSON:API Cursor Pagination (Stable, Recommended):**
- *   - page[after]: Cursor token for next page (opaque token from meta.page.endCursor)
- *   - page[before]: Cursor token for previous page (opaque token from meta.page.startCursor)
- *   - page[size]: Items per page (default: 100, max: 500)
- *   - Cursors provide stable results even when data changes during browsing
- *
- * **Legacy Parameters (Deprecated, auto-converted to JSON:API):**
- *   - skip: Number of records to skip (converted to page[number])
- *   - limit: Max records to return (converted to page[size])
- *   - sex: Filter by sex (converted to filter[sex])
- *   - has_variants: Filter by variant presence (converted to filter[has_variants])
- *
- * **Response Format (JSON:API):**
- * Returns { data, meta, links } where:
- *   - data: Array of phenopacket documents
- *   - meta.page: Offset pagination → { currentPage, pageSize, totalPages, totalRecords }
- *   - meta.page: Cursor pagination → { pageSize, hasNextPage, hasPreviousPage, startCursor, endCursor }
- *   - links: { self, first, prev, next, last } (or { self, first, prev, next } for cursor)
- *
- * @returns {Promise} Axios promise resolving to JSON:API response
- *
- * @example
- * // Offset pagination (simple, may skip/duplicate records if data changes)
- * getPhenopackets({
- *   'page[number]': 1,
- *   'page[size]': 20,
- *   'filter[sex]': 'MALE',
- *   'sort': '-created_at'
- * })
- *
- * @example
- * // Cursor pagination (stable, recommended for browsing)
- * const response1 = await getPhenopackets({ 'page[size]': 20 })
- * const nextCursor = response1.meta.page.endCursor
- * const response2 = await getPhenopackets({ 'page[after]': nextCursor, 'page[size]': 20 })
- *
- * @example
- * // Legacy style (auto-converted to JSON:API)
- * getPhenopackets({ skip: 0, limit: 20, sex: 'MALE' })
- */
-export const getPhenopackets = (params) => apiClient.get('/phenopackets/', { params });
-
-/**
- * Get a single phenopacket by ID.
- * @param {string} id - Phenopacket ID
- * @returns {Promise} Axios promise with phenopacket document
- */
-export const getPhenopacket = (id) => apiClient.get(`/phenopackets/${id}`);
-
-/**
- * Get phenotype timeline for an individual phenopacket.
- * Returns temporal data for phenotypic features with onset ages and evidence.
- * @param {string} id - Phenopacket ID
- * @returns {Promise} Axios promise with timeline data
- *   - subject_id: Subject identifier
- *   - current_age: Current age (ISO8601 duration)
- *   - features: Array of phenotypic features with onset/evidence
- */
-export const getPhenotypeTimeline = (id) => apiClient.get(`/phenopackets/${id}/timeline`);
-
-/**
- * Create a new phenopacket (requires curator role).
- * @param {Object} phenopacketData - Phenopacket data following GA4GH Phenopackets v2 standard
- * @returns {Promise} Axios promise with created phenopacket
- */
-export const createPhenopacket = (phenopacketData) =>
-  apiClient.post('/phenopackets/', phenopacketData);
-
-/**
- * Update an existing phenopacket (requires curator role).
- * @param {string} id - Phenopacket ID
- * @param {Object} data - Update data object
- * @param {Object} data.phenopacket - Updated phenopacket data
- * @param {number} data.revision - Current revision for optimistic locking
- * @param {string} data.change_reason - Reason for the change (audit trail)
- * @returns {Promise} Axios promise with updated phenopacket
- */
-export const updatePhenopacket = (id, data) =>
-  apiClient.put(`/phenopackets/${id}`, {
-    phenopacket: data.phenopacket,
-    revision: data.revision,
-    change_reason: data.change_reason,
-  });
-
-/**
- * Delete a phenopacket (soft delete, requires curator role).
- * @param {string} id - Phenopacket ID
- * @param {string} changeReason - Reason for deletion (required for audit trail)
- * @returns {Promise} Axios promise with deletion confirmation
- */
-export const deletePhenopacket = (id, changeReason) =>
-  apiClient.delete(`/phenopackets/${id}`, {
-    params: { change_reason: changeReason },
-  });
-
-/**
- * Get audit history for a phenopacket.
- * @param {string} id - Phenopacket ID
- * @returns {Promise} Axios promise with array of audit entries
- */
-export const getPhenopacketAuditHistory = (id) => apiClient.get(`/phenopackets/${id}/audit`);
-
-/**
- * Get multiple phenopackets by IDs in a single request.
- * Prevents N+1 query problem.
- * @param {Array<string>} phenopacketIds - Array of phenopacket IDs
- * @returns {Promise} Axios promise with array of phenopacket documents
- */
-export const getPhenopacketsBatch = (phenopacketIds) =>
-  apiClient.get('/phenopackets/batch', {
-    params: { phenopacket_ids: phenopacketIds.join(',') },
-  });
-
-/**
- * Search phenopackets with advanced filters using GET request.
- * @param {Object} params - Search criteria as query parameters
- *   - q: Full-text search query (optional)
- *   - hpo_id: HPO term ID (optional)
- *   - sex: Subject sex (optional)
- *   - gene: Gene symbol (optional)
- *   - pmid: Publication PMID (optional)
- *   - rank_by_relevance: Sort by search rank (optional, default: true)
- * @returns {Promise} Axios promise with search results
- */
-export const searchPhenopackets = (params) => apiClient.get('/phenopackets/search', { params });
-
-/**
- * Get facet counts for search filters based on current search criteria.
- * @param {Object} params - Search criteria as query parameters
- *   - q: Full-text search query (optional)
- *   - hpo_id: HPO term ID (optional)
- *   - sex: Subject sex (optional)
- *   - gene: Gene symbol (optional)
- *   - pmid: Publication PMID (optional)
- * @returns {Promise} Axios promise with facet counts for filters
- */
-export const getSearchFacets = (params) => apiClient.get('/phenopackets/search/facets', { params });
-
-/**
- * Get phenopackets filtered by sex.
- * @param {string} sex - Sex value (MALE, FEMALE, OTHER_SEX, UNKNOWN_SEX)
- * @param {Object} params - Additional query parameters
- * @returns {Promise} Axios promise with filtered phenopackets
- */
-export const getPhenopacketsBySex = (sex, params = {}) =>
-  apiClient.get('/phenopackets/', { params: { ...params, sex } });
-
-/**
- * Get phenopackets that contain genomic interpretations.
- * @param {Object} params - Query parameters
- * @returns {Promise} Axios promise with phenopackets containing variants
- */
-export const getPhenopacketsWithVariants = (params = {}) =>
-  apiClient.get('/phenopackets/', { params: { ...params, has_variants: true } });
-
-/**
- * Get phenopackets citing a specific publication.
- * Server-side filtering for better performance.
- * @param {string} pmid - PubMed ID (with or without PMID: prefix)
- * @param {Object} params - Query parameters
- *   - skip: Pagination offset (default: 0)
- *   - limit: Max records (default: 100, max: 500)
- *   - sex: Filter by sex (optional)
- *   - has_variants: Filter by variant presence (optional)
- * @returns {Promise} Axios promise with filtered phenopackets and pagination info
- */
-export const getPhenopacketsByPublication = (pmid, params = {}) =>
-  apiClient.get(`/phenopackets/by-publication/${pmid}`, { params });
-
-/**
- * Get phenotypic features for multiple phenopackets (batch).
- * @param {string} phenopacketIds - Comma-separated phenopacket IDs
- * @returns {Promise} Axios promise with features data
- */
-export const getPhenotypicFeaturesBatch = (phenopacketIds) =>
-  apiClient.get('/phenopackets/features/batch', {
-    params: { phenopacket_ids: phenopacketIds },
-  });
-
-/**
- * Get variants for multiple phenopackets (batch).
- * @param {string} phenopacketIds - Comma-separated phenopacket IDs
- * @returns {Promise} Axios promise with variants data
- */
-export const getVariantsBatch = (phenopacketIds) =>
-  apiClient.get('/phenopackets/variants/batch', {
-    params: { phenopacket_ids: phenopacketIds },
-  });
-
-/**
- * Get all phenopackets that contain a specific variant.
- * @param {string} variantId - The variant ID to search for
- * @returns {Promise} Axios promise with phenopackets containing this variant
- */
-export const getPhenopacketsByVariant = (variantId) =>
-  apiClient.get(`/phenopackets/by-variant/${encodeURIComponent(variantId)}`);
-
-/* ==================== AGGREGATION ENDPOINTS ==================== */
-
-/**
- * Get summary statistics.
- * @returns {Promise} Axios promise with summary stats
- */
-export const getSummaryStats = () => apiClient.get('/phenopackets/aggregate/summary');
-
-/**
- * Get sex distribution across all phenopackets.
- * @returns {Promise} Axios promise with sex distribution data
- */
-export const getSexDistribution = () => apiClient.get('/phenopackets/aggregate/sex-distribution');
-
-/**
- * Get HPO term frequency aggregation.
- * @param {Object} params - Query parameters
- *   - limit: Max number of results (default: 20)
- * @returns {Promise} Axios promise with phenotypic features aggregation
- */
-export const getPhenotypicFeaturesAggregation = (params = {}) =>
-  apiClient.get('/phenopackets/aggregate/by-feature', { params });
-
-/**
- * Get disease frequency aggregation.
- * @param {Object} params - Query parameters
- *   - limit: Max number of results (default: 20)
- * @returns {Promise} Axios promise with disease aggregation
- */
-export const getDiseaseAggregation = (params = {}) =>
-  apiClient.get('/phenopackets/aggregate/by-disease', { params });
-
-/**
- * Get variant pathogenicity distribution.
- * @param {Object} params - Query parameters
- * @param {string} params.count_mode - Count mode: 'all' (default) or 'unique'
- * @returns {Promise} Axios promise with pathogenicity counts
- */
-export const getVariantPathogenicity = (params = {}) =>
-  apiClient.get('/phenopackets/aggregate/variant-pathogenicity', { params });
-
-/**
- * Get kidney disease stage distribution.
- * @returns {Promise} Axios promise with kidney stage data
- */
-export const getKidneyStages = () => apiClient.get('/phenopackets/aggregate/kidney-stages');
-
-/**
- * Get variant type distribution (SNV, CNV, etc.).
- * @param {Object} params - Query parameters
- * @param {string} params.count_mode - Count mode: 'all' (default) or 'unique'
- * @returns {Promise} Axios promise with variant type counts
- */
-export const getVariantTypes = (params = {}) =>
-  apiClient.get('/phenopackets/aggregate/variant-types', { params });
-
-/**
- * Get publications by type (for line chart visualization).
- * Returns publications with PMID, type, phenopacket count, and year from cached metadata.
- * Note: Requires `make publications-sync` to be run first to populate year data.
- * @returns {Promise} Axios promise with publications array including year
- */
-export const getPublicationsByType = () =>
-  apiClient.get('/phenopackets/aggregate/publications-by-type');
-
-/**
- * Get pre-aggregated timeline data for publications chart.
- * Returns data aggregated by year and publication type, ready for visualization.
- * This avoids the need to query PubMed API from the frontend.
- * Note: Requires `make publications-sync` to be run first to populate year data.
- * @returns {Promise} Axios promise with timeline data array
- *   - year: Publication year
- *   - publication_type: Type of publication (case_series, research, etc.)
- *   - publication_count: Number of publications in this year/type
- *   - phenopacket_count: Total phenopackets from these publications
- */
-export const getPublicationsTimelineData = () =>
-  apiClient.get('/phenopackets/aggregate/publications-timeline-data');
-
-// =============================================================================
-// Admin API Functions (Requires Admin Authentication)
-// =============================================================================
-
-/**
- * Get admin system status and sync statistics.
- * Requires admin authentication.
- * @returns {Promise} Axios promise with system status
- */
-export const getAdminStatus = () => apiClient.get('/admin/status');
-
-/**
- * Get detailed database statistics.
- * Requires admin authentication.
- * @returns {Promise} Axios promise with statistics
- */
-export const getAdminStatistics = () => apiClient.get('/admin/statistics');
-
-/**
- * Start publication metadata sync task.
- * Requires admin authentication.
- * @param {boolean} force - If true, re-fetch all publications even if already synced
- * @returns {Promise} Axios promise with task info
- */
-export const startPublicationSync = (force = false) =>
-  apiClient.post('/admin/sync/publications', null, { params: force ? { force: true } : {} });
-
-/**
- * Get publication sync task progress.
- * Requires admin authentication.
- * @param {string} taskId - Optional task ID
- * @returns {Promise} Axios promise with progress info
- */
-export const getPublicationSyncStatus = (taskId = null) =>
-  apiClient.get('/admin/sync/publications/status', {
-    params: taskId ? { task_id: taskId } : {},
-  });
-
-/**
- * Start variant annotation sync task.
- * Requires admin authentication.
- * @param {boolean} force - If true, re-fetch all variant annotations even if already cached
- * @returns {Promise} Axios promise with task info
- */
-export const startVariantSync = (force = false) =>
-  apiClient.post('/admin/sync/variants', null, { params: force ? { force: true } : {} });
-
-/**
- * Get variant sync task progress.
- * Requires admin authentication.
- * @param {string} taskId - Optional task ID
- * @returns {Promise} Axios promise with progress info
- */
-export const getVariantSyncStatus = (taskId = null) =>
-  apiClient.get('/admin/sync/variants/status', {
-    params: taskId ? { task_id: taskId } : {},
-  });
-
-/**
- * Initialize reference data (GRCh38 + HNF1B gene + transcript + domains).
- * Requires admin authentication.
- * @returns {Promise} Axios promise with task info
- */
-export const startReferenceInit = () => apiClient.post('/admin/sync/reference/init');
-
-/**
- * Start chr17q12 genes sync from Ensembl REST API.
- * Requires admin authentication.
- * @param {boolean} force - If true, re-sync all genes
- * @returns {Promise} Axios promise with task info
- */
-export const startGenesSync = (force = false) =>
-  apiClient.post('/admin/sync/genes', null, { params: force ? { force: true } : {} });
-
-/**
- * Get chr17q12 genes sync task progress.
- * Requires admin authentication.
- * @param {string} taskId - Optional task ID
- * @returns {Promise} Axios promise with progress info
- */
-export const getGenesSyncStatus = (taskId = null) =>
-  apiClient.get('/admin/sync/genes/status', {
-    params: taskId ? { task_id: taskId } : {},
-  });
-
-/**
- * Get reference data status.
- * Requires admin authentication.
- * @returns {Promise} Axios promise with reference data status
- */
-export const getReferenceDataStatus = () => apiClient.get('/admin/reference/status');
-
-/**
- * Get publication type distribution (for donut chart).
- * Returns aggregated counts by publication type (case_series, research, case_report, etc.).
- * @returns {Promise} Axios promise with publication type aggregation
- */
-export const getPublicationTypes = () => apiClient.get('/phenopackets/aggregate/publication-types');
-
-/**
- * Get age of onset distribution.
- * @returns {Promise} Axios promise with age of onset data
- */
-export const getAgeOfOnsetAggregation = () => apiClient.get('/phenopackets/aggregate/age-of-onset');
-
-/**
- * Get survival analysis data (Kaplan-Meier curves).
- * @param {Object} params - Query parameters
- * @param {string} params.comparison - Comparison type ('variant_type', 'pathogenicity', 'disease_subtype')
- * @returns {Promise} Axios promise with survival data including groups and statistical tests
- */
-export const getSurvivalData = (params = {}) =>
-  apiClient.get('/phenopackets/aggregate/survival-data', { params });
-
-/**
- * Compare phenotype distributions between variant type groups.
- * @param {Object} params - Query parameters
- * @param {string} params.comparison - Type of comparison ('truncating_vs_non_truncating' or 'cnv_vs_point_mutation')
- * @param {number} params.limit - Maximum number of phenotypes to return (default: 20, max: 100)
- * @param {number} params.min_prevalence - Minimum prevalence (0-1) in at least one group (default: 0.05)
- * @param {string} params.sort_by - Sort by 'p_value', 'effect_size', or 'prevalence_diff' (default: 'p_value')
- * @returns {Promise} Axios promise with comparison results
- */
-export const compareVariantTypes = (params) =>
-  apiClient.get('/phenopackets/compare/variant-types', { params });
-
-/**
- * Get small variants (SNVs) for protein plot visualization.
- * @returns {Promise} Axios promise with small variants data
- */
-export const getSmallVariants = () => apiClient.get('/phenopackets/variants/small-variants');
-
-/* ==================== PUBLICATION ENDPOINTS ==================== */
-
-/**
- * Get a list of publications with JSON:API pagination, filtering, and search.
- *
- * @param {Object} params - Query parameters
- *   - page[number]: Page number (1-indexed, default: 1)
- *   - page[size]: Items per page (default: 20, max: 100)
- *   - filter[year]: Filter by publication year
- *   - filter[year_gte]: Filter by minimum year
- *   - filter[year_lte]: Filter by maximum year
- *   - filter[has_doi]: Filter by DOI presence
- *   - sort: Sort field (prefix with '-' for descending, default: '-phenopacket_count')
- *   - q: Full-text search query (searches title, authors, PMID, DOI)
- * @returns {Promise} Axios promise with JSON:API response { data, meta, links }
- *
- * @example
- * getPublications({
- *   'page[number]': 1,
- *   'page[size]': 20,
- *   'sort': '-phenopacket_count',
- *   'q': 'HNF1B'
- * })
- */
-export const getPublications = (params = {}) => apiClient.get('/publications/', { params });
-
-/**
- * Get publication metadata by PMID.
- * Fetches from PubMed API with database caching (90-day TTL).
- * @param {string} pmid - PubMed ID (with or without PMID: prefix)
- * @returns {Promise} Axios promise with publication metadata
- *   - pmid: PubMed ID
- *   - title: Article title
- *   - authors: Array of author objects {name, affiliation}
- *   - journal: Journal name
- *   - year: Publication year
- *   - doi: DOI identifier
- *   - abstract: Article abstract
- *   - data_source: "PubMed"
- *   - fetched_at: ISO timestamp
- */
-export const getPublicationMetadata = (pmid) => apiClient.get(`/publications/${pmid}/metadata`);
-
-/* ==================== AUTHENTICATION ENDPOINTS ==================== */
-
-/**
- * Login user and get JWT token.
- * @param {Object} credentials - User credentials
- *   - username: Username
- *   - password: Password
- * @returns {Promise} Axios promise with token response
- */
-export const login = (credentials) => apiClient.post('/auth/login', credentials);
-
-/**
- * Get current user information.
- * @returns {Promise} Axios promise with user data
- */
-export const getCurrentUser = () => apiClient.get('/auth/me');
-
-/**
- * Logout user (client-side token removal).
- * @returns {Promise} Axios promise with logout confirmation
- */
-export const logout = () => apiClient.post('/auth/logout');
-
-/* ==================== HPO AUTOCOMPLETE ==================== */
-
-/**
- * Get HPO term suggestions for autocomplete.
- * @param {string} query - Search query
- * @param {number} limit - Max results
- * @returns {Promise} Axios promise with HPO term suggestions
- */
-export const getHPOAutocomplete = (query, limit = 10) => {
-  return apiClient.get('/ontology/hpo/autocomplete', {
-    params: { q: query, limit },
-  });
-};
-
-/**
- * Search HPO terms by query string.
- * @param {string} query - Search query
- * @returns {Promise} Axios promise with HPO term suggestions
- */
-export const searchHPOTerms = (query) =>
-  apiClient.get('/hpo/autocomplete', { params: { q: query } });
-
-/* ==================== CLINICAL ENDPOINTS ==================== */
-
-/**
- * Get phenopackets with renal insufficiency.
- * @returns {Promise} Axios promise with clinical data
- */
-export const getRenalInsufficiencyCases = () => apiClient.get('/clinical/renal-insufficiency');
-
-/**
- * Get phenopackets with genital abnormalities.
- * @returns {Promise} Axios promise with clinical data
- */
-export const getGenitalAbnormalitiesCases = () => apiClient.get('/clinical/genital-abnormalities');
-
-/**
- * Get phenopackets with diabetes.
- * @returns {Promise} Axios promise with clinical data
- */
-export const getDiabetesCases = () => apiClient.get('/clinical/diabetes');
-
-/**
- * Get phenopackets with hypomagnesemia.
- * @returns {Promise} Axios promise with clinical data
- */
-export const getHypomagnesemiaCases = () => apiClient.get('/clinical/hypomagnesemia');
-
-/**
- * Get aggregated unique variants across all phenopackets with search and filters.
- * Uses offset-based pagination (JSON:API v1.1).
- *
- * @param {Object} params - Query parameters
- * @param {number} params.page - Page number (1-indexed, default: 1)
- * @param {number} params.pageSize - Items per page (default: 10)
- * @param {string} [params.query] - Text search (HGVS, variant ID, coordinates)
- * @param {string} [params.variant_type] - Filter by variant type (SNV, deletion, etc.)
- * @param {string} [params.classification] - Filter by ACMG classification
- * @param {string} [params.consequence] - Filter by molecular consequence
- * @param {string} [params.sort] - Sort field (prefix with '-' for descending order)
- * @returns {Promise} Promise resolving to variants data with offset pagination metadata
- */
-export const getVariants = async (params = {}) => {
-  const {
-    page = 1,
-    pageSize = 10,
-    query,
-    variant_type,
-    classification,
-    consequence,
-    domain,
-    sort,
-  } = params;
-
-  const response = await apiClient.get('/phenopackets/aggregate/all-variants', {
-    params: {
-      'page[number]': page,
-      'page[size]': pageSize,
-      query,
-      variant_type,
-      classification,
-      consequence,
-      domain,
-      sort,
-    },
-  });
-
-  // Backend returns JSON:API offset response
-  const responseData = response.data || {};
-  const data = responseData.data || [];
-  const meta = responseData.meta?.page || {};
-
-  return {
-    data: data.map((variant) => ({
-      id: variant.simple_id || variant.variant_id,
-      simple_id: variant.simple_id,
-      variant_id: variant.variant_id,
-      label: variant.label,
-      geneSymbol: variant.gene_symbol,
-      geneId: variant.gene_id,
-      variant_type: variant.structural_type,
-      hg38: variant.hg38,
-      transcript: variant.transcript,
-      protein: variant.protein,
-      classificationVerdict: variant.pathogenicity,
-      pathogenicity: variant.pathogenicity, // Alias for DNA Distance Analysis
-      individualCount: variant.phenopacket_count,
-      molecular_consequence: variant.molecular_consequence,
-    })),
-    meta: {
-      currentPage: meta.currentPage || page,
-      pageSize: meta.pageSize || pageSize,
-      totalPages: meta.totalPages || 0,
-      totalRecords: meta.totalRecords || 0,
-    },
-  };
-};
-
-/* ==================== REFERENCE GENOME ENDPOINTS ==================== */
-
-/**
- * Get list of all available genome assemblies.
- * @returns {Promise} Axios promise with genome assemblies
- */
-export const getReferenceGenomes = () => apiClient.get('/reference/genomes');
-
-/**
- * Query genes by symbol or chromosome.
- * @param {Object} params - Query parameters
- *   - symbol: Gene symbol to filter (e.g., "HNF1B")
- *   - chromosome: Chromosome to filter (e.g., "17")
- *   - genome_build: Genome assembly name (default: GRCh38)
- * @returns {Promise} Axios promise with genes array
- */
-export const getReferenceGenes = (params = {}) => apiClient.get('/reference/genes', { params });
-
-/**
- * Get gene details with transcripts.
- * @param {string} symbol - Gene symbol (e.g., "HNF1B")
- * @param {string} genomeBuild - Genome assembly name (default: GRCh38)
- * @returns {Promise} Axios promise with gene details
- */
-export const getReferenceGene = (symbol, genomeBuild = 'GRCh38') =>
-  apiClient.get(`/reference/genes/${symbol}`, {
-    params: { genome_build: genomeBuild },
-  });
-
-/**
- * Get all transcript isoforms for a gene with exon coordinates.
- * @param {string} symbol - Gene symbol (e.g., "HNF1B")
- * @param {string} genomeBuild - Genome assembly name (default: GRCh38)
- * @returns {Promise} Axios promise with transcripts array
- */
-export const getReferenceGeneTranscripts = (symbol, genomeBuild = 'GRCh38') =>
-  apiClient.get(`/reference/genes/${symbol}/transcripts`, {
-    params: { genome_build: genomeBuild },
-  });
-
-/**
- * Get protein domains for a gene's canonical transcript.
- * Returns empty domains array gracefully if reference data is unavailable.
- * @param {string} symbol - Gene symbol (e.g., "HNF1B")
- * @param {string} genomeBuild - Genome assembly name (default: GRCh38)
- * @returns {Promise} Axios promise with protein domains
- *   - gene: Gene symbol
- *   - protein: RefSeq protein ID
- *   - uniprot: UniProt accession
- *   - length: Protein length (amino acids)
- *   - domains: Array of domain objects with name, start, end, function
- *   - genome_build: Genome assembly
- *   - updated_at: Last update timestamp
- */
-export const getReferenceGeneDomains = async (symbol, genomeBuild = 'GRCh38') => {
-  try {
-    return await apiClient.get(`/reference/genes/${symbol}/domains`, {
-      params: { genome_build: genomeBuild },
-    });
-  } catch (error) {
-    // Handle 404 gracefully - reference data may not be populated
-    if (error.response?.status === 404) {
-      window.logService?.debug('Reference domains not available', { symbol, genomeBuild });
-      return { data: { domains: [], gene: symbol, genome_build: genomeBuild } };
-    }
-    throw error;
-  }
-};
-
-/**
- * Get all genes in a genomic region.
- * @param {string} region - Genomic region in format "chr:start-end" (e.g., "17:36000000-37000000")
- * @param {string} genomeBuild - Genome assembly name (default: GRCh38)
- * @returns {Promise} Axios promise with genes in region
- */
-export const getReferenceGenomicRegion = (region, genomeBuild = 'GRCh38') =>
-  apiClient.get(`/reference/regions/${region}`, {
-    params: { genome_build: genomeBuild },
-  });
-
-/* ==================== VARIANT ANNOTATION ENDPOINTS ==================== */
-
-/**
- * Annotate a variant using Ensembl Variant Effect Predictor (VEP).
- * Returns comprehensive variant annotations including consequence predictions,
- * impact severity, CADD scores, and gnomAD frequencies.
- *
- * @param {string} variant - Variant notation in one of these formats:
- *   - HGVS: "NM_000458.4:c.544+1G>A" or "NC_000017.11:g.36459258A>G"
- *   - VCF: "17-36459258-A-G" or "chr17-36459258-A-G"
- *   - rsID: "rs56116432"
- * @returns {Promise} Axios promise with VEP annotation data
- *   - id: Variant identifier
- *   - input: Original input notation
- *   - allele_string: Reference/alternate alleles
- *   - most_severe_consequence: Most severe predicted consequence (e.g., "missense_variant")
- *   - transcript_consequences: Array of transcript annotations
- *   - colocated_variants: Array of known variants (rsIDs, gnomAD)
- *   - cadd: CADD scores object (PHRED, raw) if available
- *   - gnomad: gnomAD allele frequency object if available
- *   - impact: Impact severity (HIGH, MODERATE, LOW, MODIFIER)
- */
-export const annotateVariant = (variant) =>
-  apiClient.post('/variants/annotate', null, {
-    params: { variant },
-  });
-
-/* ==================== GLOBAL SEARCH ENDPOINTS ==================== */
-
-/**
- * Get autocomplete suggestions for global search.
- * @param {string} q - Search query
- * @param {number} limit - Max results
- * @returns {Promise} Axios promise with search suggestions
- */
-export const searchAutocomplete = (q, limit = 10) =>
-  apiClient.get('/search/autocomplete', { params: { q, limit } });
-
-/**
- * Perform a global full-text search.
- * @param {string} q - Search query
- * @param {number} page - Page number
- * @param {number} pageSize - Page size
- * @param {string} type - Filter by type
- * @returns {Promise} Axios promise with search results
- */
-export const searchGlobal = (q, page = 1, pageSize = 20, type = null) =>
-  apiClient.get('/search/global', { params: { q, page, page_size: pageSize, type } });
-
-// Default export with all API methods
 export default {
-  // Phenopackets CRUD
-  getPhenopackets,
-  getPhenopacket,
-  getPhenotypeTimeline,
-  createPhenopacket,
-  updatePhenopacket,
-  deletePhenopacket,
-  getPhenopacketAuditHistory,
-
-  // Phenopackets Batch & Search
-  getPhenopacketsBatch,
-  searchPhenopackets,
-  getSearchFacets,
-  getPhenopacketsBySex,
-  getPhenopacketsWithVariants,
-  getPhenopacketsByPublication,
-  getPhenotypicFeaturesBatch,
-  getVariantsBatch,
-  getPhenopacketsByVariant,
-
-  // Aggregations
-  getSummaryStats,
-  getSexDistribution,
-  getPhenotypicFeaturesAggregation,
-  getDiseaseAggregation,
-  getVariantPathogenicity,
-  getKidneyStages,
-  getVariantTypes,
-  getPublicationsByType,
-  getPublicationsTimelineData,
-  getPublicationTypes,
-  getAgeOfOnsetAggregation,
-  getSurvivalData,
-  getSmallVariants,
-
-  // Publications
-  getPublications,
-  getPublicationMetadata,
-
-  // Authentication
-  login,
-  getCurrentUser,
-  logout,
-
-  // HPO
-  getHPOAutocomplete,
-  searchHPOTerms,
-
-  // Clinical endpoints
-  getRenalInsufficiencyCases,
-  getGenitalAbnormalitiesCases,
-  getDiabetesCases,
-  getHypomagnesemiaCases,
-
-  // Variants
-  getVariants,
-
-  // Reference
-  getReferenceGenomes,
-  getReferenceGenes,
-  getReferenceGene,
-  getReferenceGeneTranscripts,
-  getReferenceGeneDomains,
-  getReferenceGenomicRegion,
-
-  // Variant annotation
-  annotateVariant,
-
-  // Global Search
-  searchAutocomplete,
-  searchGlobal,
-
-  // Admin (requires admin authentication)
-  getAdminStatus,
-  getAdminStatistics,
-  startPublicationSync,
-  getPublicationSyncStatus,
-  startVariantSync,
-  getVariantSyncStatus,
-  startReferenceInit,
-  startGenesSync,
-  getGenesSyncStatus,
-  getReferenceDataStatus,
-
-  // Axios client for custom requests
+  ...phenopackets,
+  ...aggregations,
+  ...admin,
+  ...publications,
+  ...auth,
+  ...hpo,
+  ...clinical,
+  ...variants,
+  ...reference,
+  ...variantAnnotation,
+  ...search,
   client: apiClient,
 };
